+++ src/Symfony/Component/Form/Form.php
@@ -89,10 +89,10 @@ class Form implements \IteratorAggregate, FormInterface
     private $bound = false;
 
     /**
-     * The form data in application format
+     * The form data in model format
      * @var mixed
      */
-    private $appData;
+    private $modelData;
 
     /**
      * The form data in normalized format
@@ -101,10 +101,10 @@ class Form implements \IteratorAggregate, FormInterface
     private $normData;
 
     /**
-     * The form data in client format
+     * The form data in view format
      * @var mixed
      */
-    private $clientData;
+    private $viewData;
 
     /**
      * The bound values that don't belong to any children
@@ -113,7 +113,7 @@ class Form implements \IteratorAggregate, FormInterface
     private $extraData = array();
 
     /**
-     * Whether the data in application, normalized and client format is
+     * Whether the data in model, normalized and view format is
      * synchronized. Data may not be synchronized if transformation errors
      * occur.
      * @var Boolean
@@ -315,75 +315,75 @@ class Form implements \IteratorAggregate, FormInterface
     /**
      * Updates the form with default data.
      *
-     * @param array $appData The data formatted as expected for the underlying object
+     * @param array $modelData The data formatted as expected for the underlying object
      *
      * @return Form The current form
      */
-    public function setData($appData)
+    public function setData($modelData)
     {
         if ($this->bound) {
             throw new AlreadyBoundException('You cannot change the data of a bound form');
         }
 
-        if (is_object($appData) && !$this->config->getByReference()) {
-            $appData = clone $appData;
+        if (is_object($modelData) && !$this->config->getByReference()) {
+            $modelData = clone $modelData;
         }
 
-        $event = new DataEvent($this, $appData);
+        $event = new DataEvent($this, $modelData);
         $this->config->getEventDispatcher()->dispatch(FormEvents::PRE_SET_DATA, $event);
 
         // Hook to change content of the data
-        $event = new FilterDataEvent($this, $appData);
+        $event = new FilterDataEvent($this, $modelData);
         $this->config->getEventDispatcher()->dispatch(FormEvents::SET_DATA, $event);
-        $appData = $event->getData();
+        $modelData = $event->getData();
 
         // Treat data as strings unless a value transformer exists
-        if (!$this->config->getClientTransformers() && !$this->config->getNormTransformers() && is_scalar($appData)) {
-            $appData = (string) $appData;
+        if (!$this->config->getViewTransformers() && !$this->config->getModelTransformers() && is_scalar($modelData)) {
+            $modelData = (string) $modelData;
         }
 
         // Synchronize representations - must not change the content!
-        $normData = $this->appToNorm($appData);
-        $clientData = $this->normToClient($normData);
+        $normData = $this->modelToNorm($modelData);
+        $viewData = $this->normToView($normData);
 
-        // Validate if client data matches data class (unless empty)
-        if (!empty($clientData)) {
+        // Validate if view data matches data class (unless empty)
+        if (!empty($viewData)) {
             $dataClass = $this->config->getDataClass();
 
-            if (null === $dataClass && is_object($clientData) && !$clientData instanceof \ArrayAccess) {
+            if (null === $dataClass && is_object($viewData) && !$viewData instanceof \ArrayAccess) {
                 $expectedType = 'scalar, array or an instance of \ArrayAccess';
 
                 throw new FormException(
-                    'The form\'s client data is expected to be of type ' . $expectedType . ', ' .
-                    'but is an instance of class ' . get_class($clientData) . '. You ' .
+                    'The form\'s view data is expected to be of type ' . $expectedType . ', ' .
+                    'but is an instance of class ' . get_class($viewData) . '. You ' .
                     'can avoid this error by setting the "data_class" option to ' .
-                    '"' . get_class($clientData) . '" or by adding a client transformer ' .
-                    'that transforms ' . get_class($clientData) . ' to ' . $expectedType . '.'
+                    '"' . get_class($viewData) . '" or by adding a view transformer ' .
+                    'that transforms ' . get_class($viewData) . ' to ' . $expectedType . '.'
                 );
             }
 
-            if (null !== $dataClass && !$clientData instanceof $dataClass) {
+            if (null !== $dataClass && !$viewData instanceof $dataClass) {
                 throw new FormException(
-                    'The form\'s client data is expected to be an instance of class ' .
-                    $dataClass . ', but has the type ' . gettype($clientData) . '. You ' .
+                    'The form\'s view data is expected to be an instance of class ' .
+                    $dataClass . ', but has the type ' . gettype($viewData) . '. You ' .
                     'can avoid this error by setting the "data_class" option to ' .
-                    'null or by adding a client transformer that transforms ' .
-                    gettype($clientData) . ' to ' . $dataClass . '.'
+                    'null or by adding a view transformer that transforms ' .
+                    gettype($viewData) . ' to ' . $dataClass . '.'
                 );
             }
         }
 
-        $this->appData = $appData;
+        $this->modelData = $modelData;
         $this->normData = $normData;
-        $this->clientData = $clientData;
+        $this->viewData = $viewData;
         $this->synchronized = true;
 
         if (count($this->children) > 0 && $this->config->getDataMapper()) {
             // Update child forms from the data
-            $this->config->getDataMapper()->mapDataToForms($clientData, $this->children);
+            $this->config->getDataMapper()->mapDataToForms($viewData, $this->children);
         }
 
-        $event = new DataEvent($this, $appData);
+        $event = new DataEvent($this, $modelData);
         $this->config->getEventDispatcher()->dispatch(FormEvents::POST_SET_DATA, $event);
 
         return $this;
@@ -396,7 +396,7 @@ class Form implements \IteratorAggregate, FormInterface
      */
     public function getData()
     {
-        return $this->appData;
+        return $this->modelData;
     }
 
     /**
@@ -404,9 +404,22 @@ class Form implements \IteratorAggregate, FormInterface
      *
      * @return string
      */
+    public function getViewData()
+    {
+        return $this->viewData;
+    }
+
+    /**
+     * Alias of {@link getViewData()}.
+     *
+     * @return string
+     *
+     * @deprecated Deprecated since version 2.1, to be removed in 2.3. Use
+     *             {@link getViewData()} instead.
+     */
     public function getClientData()
     {
-        return $this->clientData;
+        return $this->getViewData();
     }
 
     /**
@@ -422,13 +435,13 @@ class Form implements \IteratorAggregate, FormInterface
     /**
      * Binds data to the form, transforms and validates it.
      *
-     * @param string|array $clientData The data
+     * @param string|array $submittedData The data
      *
      * @return Form The current form
      *
      * @throws UnexpectedTypeException
      */
-    public function bind($clientData)
+    public function bind($submittedData)
     {
         if ($this->bound) {
             throw new AlreadyBoundException('A form can only be bound once');
@@ -444,43 +457,46 @@ class Form implements \IteratorAggregate, FormInterface
         // whether an empty value has been submitted or whether no value has
         // been submitted at all. This is important for processing checkboxes
         // and radio buttons with empty values.
-        if (is_scalar($clientData)) {
-            $clientData = (string) $clientData;
+        if (is_scalar($submittedData)) {
+            $submittedData = (string) $submittedData;
         }
 
         // Initialize errors in the very beginning so that we don't lose any
         // errors added during listeners
         $this->errors = array();
 
-        $event = new DataEvent($this, $clientData);
+        $event = new DataEvent($this, $submittedData);
         $this->config->getEventDispatcher()->dispatch(FormEvents::PRE_BIND, $event);
 
-        $appData = null;
+        $modelData = null;
         $normData = null;
         $extraData = array();
         $synchronized = false;
 
         // Hook to change content of the data bound by the browser
-        $event = new FilterDataEvent($this, $clientData);
+        $event = new FilterDataEvent($this, $submittedData);
         $this->config->getEventDispatcher()->dispatch(FormEvents::BIND_CLIENT_DATA, $event);
-        $clientData = $event->getData();
+        $submittedData = $event->getData();
+
+        // Build the data in the view format
+        $viewData = $submittedData;
 
         if (count($this->children) > 0) {
-            if (null === $clientData || '' === $clientData) {
-                $clientData = array();
+            if (null === $viewData || '' === $viewData) {
+                $viewData = array();
             }
 
-            if (!is_array($clientData)) {
-                throw new UnexpectedTypeException($clientData, 'array');
+            if (!is_array($viewData)) {
+                throw new UnexpectedTypeException($viewData, 'array');
             }
 
             foreach ($this->children as $name => $child) {
-                if (!isset($clientData[$name])) {
-                    $clientData[$name] = null;
+                if (!isset($viewData[$name])) {
+                    $viewData[$name] = null;
                 }
             }
 
-            foreach ($clientData as $name => $value) {
+            foreach ($viewData as $name => $value) {
                 if ($this->has($name)) {
                     $this->children[$name]->bind($value);
                 } else {
@@ -488,31 +504,32 @@ class Form implements \IteratorAggregate, FormInterface
                 }
             }
 
-            // If we have a data mapper, use old client data and merge
+            // If we have a data mapper, use old view data and merge
             // data from the children into it later
             if ($this->config->getDataMapper()) {
-                $clientData = $this->getClientData();
+                $viewData = $this->getViewData();
             }
         }
 
-        if (null === $clientData || '' === $clientData) {
+        if (null === $viewData || '' === $viewData) {
             $emptyData = $this->config->getEmptyData();
 
             if ($emptyData instanceof \Closure) {
-                $emptyData = $emptyData($this, $clientData);
+                /* @var \Closure $emptyData */
+                $emptyData = $emptyData($this, $viewData);
             }
 
-            $clientData = $emptyData;
+            $viewData = $emptyData;
         }
 
-        // Merge form data from children into existing client data
-        if (count($this->children) > 0 && $this->config->getDataMapper() && null !== $clientData) {
-            $this->config->getDataMapper()->mapFormsToData($this->children, $clientData);
+        // Merge form data from children into existing view data
+        if (count($this->children) > 0 && $this->config->getDataMapper() && null !== $viewData) {
+            $this->config->getDataMapper()->mapFormsToData($this->children, $viewData);
         }
 
         try {
             // Normalize data to unified representation
-            $normData = $this->clientToNorm($clientData);
+            $normData = $this->viewToNorm($viewData);
             $synchronized = true;
         } catch (TransformationFailedException $e) {
         }
@@ -525,18 +542,18 @@ class Form implements \IteratorAggregate, FormInterface
             $normData = $event->getData();
 
             // Synchronize representations - must not change the content!
-            $appData = $this->normToApp($normData);
-            $clientData = $this->normToClient($normData);
+            $modelData = $this->normToModel($normData);
+            $viewData = $this->normToView($normData);
         }
 
         $this->bound = true;
-        $this->appData = $appData;
+        $this->modelData = $modelData;
         $this->normData = $normData;
-        $this->clientData = $clientData;
+        $this->viewData = $viewData;
         $this->extraData = $extraData;
         $this->synchronized = $synchronized;
 
-        $event = new DataEvent($this, $clientData);
+        $event = new DataEvent($this, $viewData);
         $this->config->getEventDispatcher()->dispatch(FormEvents::POST_BIND, $event);
 
         foreach ($this->config->getValidators() as $validator) {
@@ -673,7 +690,7 @@ class Form implements \IteratorAggregate, FormInterface
             }
         }
 
-        return array() === $this->appData || null === $this->appData || '' === $this->appData;
+        return array() === $this->modelData || null === $this->modelData || '' === $this->modelData;
     }
 
     /**
@@ -761,11 +778,11 @@ class Form implements \IteratorAggregate, FormInterface
      * @return array An array of DataTransformerInterface
      *
      * @deprecated Deprecated since version 2.1, to be removed in 2.3. Use
-     *             {@link getConfig()} and {@link FormConfigInterface::getNormTransformers()} instead.
+     *             {@link getConfig()} and {@link FormConfigInterface::getModelTransformers()} instead.
      */
     public function getNormTransformers()
     {
-        return $this->config->getNormTransformers();
+        return $this->config->getModelTransformers();
     }
 
     /**
@@ -774,11 +791,11 @@ class Form implements \IteratorAggregate, FormInterface
      * @return array An array of DataTransformerInterface
      *
      * @deprecated Deprecated since version 2.1, to be removed in 2.3. Use
-     *             {@link getConfig()} and {@link FormConfigInterface::getClientTransformers()} instead.
+     *             {@link getConfig()} and {@link FormConfigInterface::getViewTransformers()} instead.
      */
     public function getClientTransformers()
     {
-        return $this->config->getClientTransformers();
+        return $this->config->getViewTransformers();
     }
 
     /**
@@ -815,7 +832,7 @@ class Form implements \IteratorAggregate, FormInterface
         $child->setParent($this);
 
         if ($this->config->getDataMapper()) {
-            $this->config->getDataMapper()->mapDataToForms($this->getClientData(), array($child));
+            $this->config->getDataMapper()->mapDataToForms($this->getViewData(), array($child));
         }
 
         return $this;
@@ -974,9 +991,9 @@ class Form implements \IteratorAggregate, FormInterface
      *
      * @return string
      */
-    private function appToNorm($value)
+    private function modelToNorm($value)
     {
-        foreach ($this->config->getNormTransformers() as $transformer) {
+        foreach ($this->config->getModelTransformers() as $transformer) {
             $value = $transformer->transform($value);
         }
 
@@ -990,9 +1007,9 @@ class Form implements \IteratorAggregate, FormInterface
      *
      * @return mixed
      */
-    private function normToApp($value)
+    private function normToModel($value)
     {
-        $transformers = $this->config->getNormTransformers();
+        $transformers = $this->config->getModelTransformers();
 
         for ($i = count($transformers) - 1; $i >= 0; --$i) {
             $value = $transformers[$i]->reverseTransform($value);
@@ -1008,15 +1025,15 @@ class Form implements \IteratorAggregate, FormInterface
      *
      * @return string
      */
-    private function normToClient($value)
+    private function normToView($value)
     {
-        if (!$this->config->getClientTransformers()) {
+        if (!$this->config->getViewTransformers()) {
             // Scalar values should always be converted to strings to
             // facilitate differentiation between empty ("") and zero (0).
             return null === $value || is_scalar($value) ? (string) $value : $value;
         }
 
-        foreach ($this->config->getClientTransformers() as $transformer) {
+        foreach ($this->config->getViewTransformers() as $transformer) {
             $value = $transformer->transform($value);
         }
 
@@ -1030,9 +1047,9 @@ class Form implements \IteratorAggregate, FormInterface
      *
      * @return mixed
      */
-    private function clientToNorm($value)
+    private function viewToNorm($value)
     {
-        $transformers = $this->config->getClientTransformers();
+        $transformers = $this->config->getViewTransformers();
 
         if (!$transformers) {
             return '' === $value ? null : $value;